Makefile: Nouveau systeme d'autoconfiguration. io.[ch]: l'argument "longueur" de getblock et putblock est un long.
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@627 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ byterun/io.h
@@ -52,13 +52,13 @@ struct channel {
 struct channel * open_descr P((int));
 value flush P((struct channel *));
 void putword P((struct channel *, uint32));
-void putblock P((struct channel *, char *, unsigned));
+void putblock P((struct channel *, char *, long));
 unsigned char refill P((struct channel *));
 value pos_out P((struct channel *));
 value seek_out P((struct channel *, value));
 uint32 getword P((struct channel *));
-unsigned getblock P((struct channel *, char *, unsigned));
-int really_getblock P((struct channel *, char *, unsigned long));
+int getblock P((struct channel *, char *, long));
+int really_getblock P((struct channel *, char *, long));
 value close_in P((struct channel *));
